Fix reading rows with skipped reference attribute
Some Excel files generated with third-party libraries such as SpreadsheetGear - there could be other, skip the "r" attribute on "c" tags within rows. Excel seems to put this attribute on every cell, but I couldn't find any evidence that the standard commands it.

openpyxl is not expecting these attributes to be skipped, so the result is that we end up with an invalid parsed row.

To fix this, when we get coordinates, make sure col_counter reflects them so the reader knows which cell it has been reading.

--HG--
branch : 3.0